+++ Class/Conf.cpp
@@ -1,0 +1,58 @@
+#include "../includes/webserv.hpp"
+#include <algorithm>
+#include <cctype>
+#include <iterator>
+#include <map>
+#include <string>
+#include <utility>
+#include <vector>
+
+conf::conf() {}
+
+void conf::check()
+{
+	if (_http.size() == 0)
+		throw exc ("Error: http not found!\n");
+	if (_servers.size() == 0)
+		throw exc ("Error: server not found!\n");
+	for (std::map<int, server>::iterator it = _servers.begin(); it != _servers.end(); ++it)
+	{
+		if (it->second.checkLocation() == 0)
+			throw exc("Error: location not found!\n");
+	}
+}
+
+
+void conf::addServer(int nbrServer, const server &srv)
+{
+	if (srv.getSize() == 0)
+		return ;
+	if (_servers.find(nbrServer) != _servers.end())
+		_servers.erase(nbrServer);
+	_servers.insert(std::make_pair(nbrServer, srv));
+}
+
+void conf::addHttp(const http& http) {
+	if (http.getSize() == 0)
+		return ;
+	if (_http.size() != 0)
+		_http.pop_back();
+	_http.push_back(http);
+}
+
+
+void conf::printServer()
+{
+	for (std::map<int,server>::iterator it = _servers.begin(); it != _servers.end(); it++) {
+		std::cout << "Key == " << it->first << '\n';
+		it->second.printMap();
+		std::cout << "---LOCATIONS---\n";
+		it->second.printLocation();
+		std::cout<<"\n";
+    }
+	for (std::vector<http>::iterator it = _http.begin(); it != _http.end(); it++) {
+		it->printMap();
+	}
+}
+
+conf::~conf() {}
