+++ Makefile
@@ -1,4 +1,6 @@
-SRC = main.cpp Class/Class.cpp Parsing/parsing.cpp
+SRC =	main.cpp \
+		Class/ABlock.cpp Class/Conf.cpp Class/Http.cpp Class/Server.cpp \
+		Parsing/Parsing.cpp
 OBJ = $(SRC:.cpp=.o)
 CC = c++
 CFLAGS = -Wall -Wextra -Werror -g -std=c++98
