+++ test.conf
@@ -1,5 +1,4 @@
 
-
 http {
 	client_max_body_size 10M;
 	client_body_timeout 60s;
@@ -57,7 +56,6 @@ http {
 		index index.html;
 		client_max_body_size 20M;
 		client_body_timeout 60s;
-		
 		autoindex on;
 		
 		location /images {
