+++ testConfRiccardo/test1.conf
@@ -1,0 +1,78 @@
+#LISTEN DUPLICATE
+http {
+	client_max_body_size 10M;
+	client_body_timeout 60s;
+
+	error_page 404 /404.html;
+	error_page 408 /408.html;
+	error_page 500 /50x.html;
+	include /etc/nginx/fastcgi_params;
+	server {
+		listen 8080;
+		listen 8081;
+		server_name localhost;
+		server_name dreamteam;
+
+		root Root;
+
+		index index.html;
+
+		client_max_body_size 10M;
+		client_body_timeout 60s;
+		autoindex off;
+
+		location /kapouet {
+			root /www;
+			index index.html;
+		}
+
+		location /kapouet/upload {
+			dav_methods GET POST DELETE;
+			root /www;
+			client_max_body_size 50M;
+			upload_pass /tmp/uploads; 
+			upload_store /tmp/uploads/;
+		}
+
+		location ~ \.php$ {
+			fastcgi_pass 127.0.0.1:9000;
+			fastcgi_index index.php;
+			fastcgi_param SCRIPT_FILENAME /tmp/www$fastcgi_script_name;
+			include fastcgi_params;
+		}
+
+		location / {
+			root /tmp/www;
+			try_files $uri $uri/ =404;
+		}
+	}
+
+	server {
+		listen 9090;
+		listen 8082;
+		server_name localhost;
+		server_name example.html;
+		root Root/example;
+		index index.html;
+		client_max_body_size 20M;
+		client_body_timeout 60s;
+		autoindex on;
+		
+		location /images {
+			root /www;  
+			index index.html; 
+		}
+
+		location ~ \.cgi$ {
+			fastcgi_pass 127.0.0.1:9000;
+			fastcgi_index index.cgi;
+			fastcgi_param SCRIPT_FILENAME /var/www/example$fastcgi_script_name;
+			include fastcgi_params;
+		}
+
+		location / {
+			root /www/example;
+			try_files $uri $uri/ =404;
+		}
+	}
+}
